refactor(web): 코드리뷰 반영
폴더 구조 수정, 안쓰이는 import문 제

JIP-101
