+++ apps/web/src/features/register/components/complete-modal/styled.tsx
