+++ apps/web/src/features/register/components/select/styled.tsx
